chore: Moved Comparator functions from Benchmark to Establishment API
Old endpoints marked as deprecated and set to return `410 Gone`
