+++ platform/src/apis/Platform.Api.Benchmark/Configuration/Services.cs
@@ -1,20 +1,15 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Text.Json;
-using Azure;
 using FluentValidation;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.DependencyInjection;
-using Platform.Api.Benchmark.Comparators;
 using Platform.Api.Benchmark.ComparatorSets;
 using Platform.Api.Benchmark.CustomData;
 using Platform.Api.Benchmark.FinancialPlans;
 using Platform.Api.Benchmark.UserData;
 using Platform.Functions.Extensions;
-using Platform.Infrastructure;
-using Platform.Search;
 using Platform.Sql;
-
 namespace Platform.Api.Benchmark.Configuration;
 
 [ExcludeFromCodeCoverage]
@@ -34,17 +29,10 @@ internal static class Services
             .AddHealthChecks()
             .AddSqlServer(sqlConnString);
 
-        var searchEndpoint = new Uri($"https://{searchName}.search.windows.net/");
-        var searchCredential = new AzureKeyCredential(searchKey);
-
         serviceCollection
             .AddSingleton<IDatabaseFactory>(new DatabaseFactory(sqlConnString))
-            .AddSingleton<ISearchConnection<ComparatorSchool>>(new SearchConnection<ComparatorSchool>(searchEndpoint, searchCredential, ResourceNames.Search.Indexes.SchoolComparators))
-            .AddSingleton<ISearchConnection<ComparatorTrust>>(new SearchConnection<ComparatorTrust>(searchEndpoint, searchCredential, ResourceNames.Search.Indexes.TrustComparators))
             .AddSingleton<IComparatorSetsService, ComparatorSetsService>()
             .AddSingleton<IFinancialPlansService, FinancialPlansService>()
-            .AddSingleton<IComparatorSchoolsService, ComparatorSchoolsService>()
-            .AddSingleton<IComparatorTrustsService, ComparatorTrustsService>()
             .AddSingleton<IUserDataService, UserDataService>()
             .AddSingleton<ICustomDataService, CustomDataService>();
