+++ platform/src/apis/Platform.Api.Benchmark/Platform.Api.Benchmark.csproj
@@ -17,23 +17,22 @@
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
             <CopyToPublishDirectory>Never</CopyToPublishDirectory>
         </None>
-        <ProjectReference Include="..\..\abstractions\Platform.Functions\Platform.Functions.csproj" />
-        <PackageReference Include="AspNetCore.HealthChecks.SqlServer" />
-        <PackageReference Include="Dapper.Contrib" />
-        <PackageReference Include="Dapper.SqlBuilder" />
-        <PackageReference Include="Microsoft.ApplicationInsights.WorkerService" />
-        <PackageReference Include="Microsoft.Azure.Functions.Worker.ApplicationInsights" />
-        <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.OpenApi" />
-        <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Storage" />
-        <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Storage.Queues" />
-        <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" />
-        <PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" />
-        <PackageReference Include="Microsoft.Azure.Functions.Worker" />
-        <ProjectReference Include="..\..\abstractions\Platform.Infrastructure\Platform.Infrastructure.csproj" />
-        <ProjectReference Include="..\..\abstractions\Platform.Search\Platform.Search.csproj" />
-        <ProjectReference Include="..\..\abstractions\Platform.Sql\Platform.Sql.csproj" />
+        <ProjectReference Include="..\..\abstractions\Platform.Functions\Platform.Functions.csproj"/>
+        <PackageReference Include="AspNetCore.HealthChecks.SqlServer"/>
+        <PackageReference Include="Dapper.Contrib"/>
+        <PackageReference Include="Dapper.SqlBuilder"/>
+        <PackageReference Include="Microsoft.ApplicationInsights.WorkerService"/>
+        <PackageReference Include="Microsoft.Azure.Functions.Worker.ApplicationInsights"/>
+        <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.OpenApi"/>
+        <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Storage"/>
+        <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Storage.Queues"/>
+        <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http"/>
+        <PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk"/>
+        <PackageReference Include="Microsoft.Azure.Functions.Worker"/>
+        <ProjectReference Include="..\..\abstractions\Platform.Infrastructure\Platform.Infrastructure.csproj"/>
+        <ProjectReference Include="..\..\abstractions\Platform.Sql\Platform.Sql.csproj"/>
     </ItemGroup>
     <ItemGroup>
-        <Using Include="System.Threading.ExecutionContext" Alias="ExecutionContext" />
+        <Using Include="System.Threading.ExecutionContext" Alias="ExecutionContext"/>
     </ItemGroup>
 </Project>
