+++ platform/src/apis/Platform.Api.Establishment/Comparators/ComparatorSchool.cs
@@ -1,5 +1,5 @@
 using System.Diagnostics.CodeAnalysis;
-namespace Platform.Api.Benchmark.Comparators;
+namespace Platform.Api.Establishment.Comparators;
 
 [ExcludeFromCodeCoverage]
 public record ComparatorSchool
