+++ platform/src/apis/Platform.Api.Establishment/Comparators/ComparatorSchools.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-namespace Platform.Api.Benchmark.Comparators;
+namespace Platform.Api.Establishment.Comparators;
 
 [ExcludeFromCodeCoverage]
 public record ComparatorSchools
