+++ platform/src/apis/Platform.Api.Establishment/Comparators/ComparatorTrust.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
-namespace Platform.Api.Benchmark.Comparators;
+namespace Platform.Api.Establishment.Comparators;
 
 [ExcludeFromCodeCoverage]
 public record ComparatorTrust
