+++ platform/tests/Platform.ApiTests/Features/BenchmarkComparators.feature
@@ -1,51 +1,11 @@
-﻿Feature: Benchmark Comparators Endpoint Testing
+Feature: Benchmark Comparators Endpoint Testing
 
     Scenario: Sending a valid comparator schools request
-        Given a valid comparator schools request for school id '990000'
+        Given a valid comparator schools request
         When I submit the comparator schools request
-        Then the comparator schools should total '234' and contain:
-          | Urn    |
-          | 990211 |
-          | 990059 |
-          | 990223 |
-          | 990070 |
-          | 990111 |
-          | 990156 |
-          | 990047 |
-          | 777042 |
-          | 990082 |
-          | 990186 |
-          | 990205 |
-          | 990183 |
-          | 990239 |
-          | 990031 |
-          | 990245 |
-          | 990180 |
-          | 990115 |
-          | 990100 |
-          | 990242 |
-          | 990155 |
-          | 990146 |
-          | 990131 |
-          | 990247 |
-          | 990008 |
-          | 990226 |
-          | 990087 |
-          | 990191 |
-          | 990012 |
-          | 990122 |
+        Then the comparator schools should return 410 Gone
 
     Scenario: Sending a valid comparator trusts request
-        Given a valid comparator trusts request for company number '10192252'
+        Given a valid comparator trusts request
         When I submit the comparator trusts request
-        Then the comparator trusts should total '25' and contain:
-          | CompanyNumber |
-          | 7185046       |
-          | 7694547       |
-          | 8053276       |
-          | 7930340       |
-          | 7452885       |
-          | 7747126       |
-          | 7695504       |
-          | 7388635       |
-          | 7703941       |
+        Then the comparator trusts should return 410 Gone
