+++ platform/tests/Platform.ApiTests/Steps/BenchmarkComparatorsSteps.cs
@@ -1,107 +1,57 @@
-﻿using System.Net;
-using System.Text;
+using System.Net;
 using FluentAssertions;
-using Platform.Api.Benchmark.Comparators;
 using Platform.ApiTests.Drivers;
-using Platform.Functions.Extensions;
-using TechTalk.SpecFlow.Assist;
-using Xunit;
-
 namespace Platform.ApiTests.Steps;
 
 [Binding]
+[Scope(Feature = "Benchmark Comparators Endpoint Testing")]
 public class BenchmarkComparatorsSteps(BenchmarkApiDriver api)
 {
     private const string ComparatorSchoolsKey = "comparator-schools";
     private const string ComparatorTrustsKey = "comparator-trusts";
 
-    [Given("a valid comparator schools request for school id '(.*)'")]
-    public void GivenAValidComparatorSchoolsRequestForSchoolId(string urn)
+    [Given("a valid comparator schools request")]
+    private void GivenAValidComparatorSchoolsRequest()
     {
-        var content = new ComparatorSchoolsRequest
-        {
-            Target = urn,
-            FinanceType = new CharacteristicList { Values = ["Maintained"] },
-            TotalPupils = new CharacteristicRange { From = 100, To = 1000 }
-        };
-
         api.CreateRequest(ComparatorSchoolsKey, new HttpRequestMessage
         {
             RequestUri = new Uri("/api/comparators/schools", UriKind.Relative),
-            Method = HttpMethod.Post,
-            Content = new StringContent(content.ToJson(), Encoding.UTF8, "application/json")
+            Method = HttpMethod.Post
         });
     }
 
     [When("I submit the comparator schools request")]
+    [When("I submit the comparator trusts request")]
     public async Task WhenISubmitTheComparatorSchoolsRequest()
     {
         await api.Send();
     }
 
-    [Then("the comparator schools should total '(.*)' and contain:")]
-    public async Task ThenTheComparatorSchoolsShouldTotalAndContain(string total, Table table)
+    [Then("the comparator schools should return 410 Gone")]
+    private void ThenTheComparatorSchoolsShouldReturnGone()
     {
-        var response = api[ComparatorSchoolsKey].Response;
+        var result = api[ComparatorSchoolsKey].Response;
 
-        response.StatusCode.Should().Be(HttpStatusCode.OK);
-
-        var content = await response.Content.ReadAsByteArrayAsync();
-        var result = content.FromJson<ComparatorSchools>();
-
-        Assert.Equal(total, result.TotalSchools.ToString());
-
-        var set = new List<dynamic>();
-        foreach (var urn in result.Schools ?? [])
-        {
-            set.Add(new { Urn = urn });
-        }
-
-        table.CompareToDynamicSet(set, false);
+        result.Should().NotBeNull();
+        result.StatusCode.Should().Be(HttpStatusCode.Gone);
     }
 
-    [Given("a valid comparator trusts request for company number '(.*)'")]
-    public void GivenAValidComparatorTrustsRequestForTrustId(string companyNumber)
+    [Given("a valid comparator trusts request")]
+    private void GivenAValidComparatorTrustsRequest()
     {
-        var content = new ComparatorTrustsRequest
-        {
-            Target = companyNumber,
-            PhasesCovered = new CharacteristicList { Values = ["Secondary"] },
-            TotalPupils = new CharacteristicRange { From = 500, To = 10000 }
-        };
-
         api.CreateRequest(ComparatorTrustsKey, new HttpRequestMessage
         {
             RequestUri = new Uri("/api/comparators/trusts", UriKind.Relative),
-            Method = HttpMethod.Post,
-            Content = new StringContent(content.ToJson(), Encoding.UTF8, "application/json")
+            Method = HttpMethod.Post
         });
     }
 
-    [When("I submit the comparator trusts request")]
-    public async Task WhenISubmitTheComparatorTrustsRequest()
+    [Then("the comparator trusts should return 410 Gone")]
+    private void ThenTheComparatorTrustsShouldReturnGone()
     {
-        await api.Send();
-    }
+        var result = api[ComparatorTrustsKey].Response;
 
-    [Then("the comparator trusts should total '(.*)' and contain:")]
-    public async Task ThenTheComparatorTrustsShouldTotalAndContain(string total, Table table)
-    {
-        var response = api[ComparatorTrustsKey].Response;
-
-        response.StatusCode.Should().Be(HttpStatusCode.OK);
-
-        var content = await response.Content.ReadAsByteArrayAsync();
-        var result = content.FromJson<ComparatorTrusts>();
-
-        Assert.Equal(total, result.TotalTrusts.ToString());
-
-        var set = new List<dynamic>();
-        foreach (var companyNumber in result.Trusts ?? [])
-        {
-            set.Add(new { CompanyNumber = companyNumber });
-        }
-
-        table.CompareToDynamicSet(set, false);
+        result.Should().NotBeNull();
+        result.StatusCode.Should().Be(HttpStatusCode.Gone);
     }
 }
